refactor: split sshnp_impl by code path
