+++ packages/noports_core/lib/src/sshnp/sshnp_mixins.dart
@@ -1,0 +1,12 @@
+import 'package:noports_core/src/sshnp/sshnp_impl/sshnp_impl.dart';
+import 'package:noports_core/sshnp.dart';
+
+mixin LegacyReverseSSH on SSHNPImpl {
+
+}
+
+mixin ReverseSSH on SSHNP {
+
+}
+
+mixin DirectSSH on SSHNP {}
